qgltf: Fix MSVC warning about narrowing conversion
Use int, fixing:
qgltf.cpp(158): warning C4838: conversion from 'QFlags<QIODevice::OpenModeFlag>' to 'ushort' requires a narrowing conversion
qgltf.cpp(159): warning C4838: conversion from 'QFlags<QIODevice::OpenModeFlag>' to 'ushort' requires a narrowing conversion
qgltf.cpp(160): warning C4838: conversion from 'QFlags<QIODevice::OpenModeFlag>' to 'ushort' requires a narrowing conversion
qgltf.cpp(161): warning C4838: conversion from 'QFlags<QIODevice::OpenModeFlag>' to 'ushort' requires a narrowing conversion
qgltf.cpp(163): warning C4838: conversion from 'QFlags<QIODevice::OpenModeFlag>' to 'ushort' requires a narrowing conversion
qgltf.cpp(165): warning C4838: conversion from 'QFlags<QIODevice::OpenModeFlag>' to 'ushort' requires a narrowing conversion

+++ tools/qgltf/qgltf.cpp
@@ -151,7 +151,7 @@ static QIODevice::OpenMode openModeFromText(const char *name) noexcept
 {
     static const struct OpenModeMapping {
         char name[2];
-        ushort mode;
+        int mode;
     } openModeMapping[] = {
         { { 'r',   0 },  QIODevice::ReadOnly  },
         { { 'r', '+' },  QIODevice::ReadWrite },
